Cleaned up quite a bit of code.  Removed threading from main app. Hopefully fixed crash: [xcb] Unknown request in queue while dequeuing [xcb] Most likely this is a multi-threaded client and XInitThreads has not been called Fixed lod bug Removed the f hotkey and added some others
git-svn-id: svn+ssh://svn.pointclouds.org/pcl/trunk@8012 a9d63959-f2ad-4865-b262-bf0e56cfafb6

+++ outofcore/include/pcl/outofcore/outofcore_iterator_base.h
@@ -115,15 +115,15 @@ namespace pcl
         {
           currentNode_ = static_cast<OctreeDiskNode*> (octree_.getRootNode ());
           currentOctreeDepth_ = 0;
-          max_depth_ = octree_.getDepth ();
+          max_depth_ = static_cast<unsigned int> (octree_.getDepth ());
         }
 
         inline void
         setMaxDepth (unsigned int max_depth)
         {
-          if (max_depth > octree_.getDepth ())
+          if (max_depth > static_cast<unsigned int> (octree_.getDepth ()))
           {
-            max_depth = octree_.getDepth ();
+            max_depth = static_cast<unsigned int> (octree_.getDepth ());
           }
 
           max_depth_ = max_depth;
